+++ uefi/src/proto/hii/mod.rs
@@ -1,0 +1,5 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+//! HII Protocols
+
+pub mod config;