+++ uefi/src/proto/mod.rs
@@ -37,6 +37,7 @@ pub mod console;
 pub mod debug;
 pub mod device_path;
 pub mod driver;
+pub mod hii;
 pub mod loaded_image;
 pub mod media;
 pub mod misc;