+++ uefi/CHANGELOG.md
@@ -4,6 +4,8 @@
 - Added `ConfigTableEntry::MEMORY_ATTRIBUTES_GUID` and `ConfigTableEntry::IMAGE_SECURITY_DATABASE_GUID`.
 - Added `proto::usb::io::UsbIo`.
 - Added `proto::pci::PciRootBridgeIo`.
+- Added `proto::hii::config::ConfigKeywordHandler`.
+- Added `proto::hii::config::HiiConfigAccess`.
 
 ## Changed
 - **Breaking:** `boot::stall` now take `core::time::Duration` instead of `usize`.